refactor(model-switcher): restructure ModelSwitcher component into modular files
- Deleted the original ModelSwitcher implementation and replaced it with a modular structure.
- Introduced new files: ModelRow, ModelBadge, ProviderLogo, TriggerContent, and utils for better organization and maintainability.
- Enhanced the ModelSwitcher to utilize these components, improving readability and separation of concerns.
- Implemented a neutral skeleton loading state for improved user experience during hydration.
